requirements.txt
langchain
langchain-ollama
langchain-core
ollama
chromadb
sentence-transformers
python-dotenv